+++ src/types/common.ts
@@ -1,19 +1,37 @@
+import { components } from '@blockfrost/openapi';
+
+// generic types
+
 export type Order = 'asc' | 'desc';
 export type AddressType = 'byron' | 'shelley';
-export interface Amount {
-  unit: string;
-  quantity: string;
-}
 
-export interface AmountExtended extends Amount {
-  decimals: null | number;
-  has_nft_onchain_metadata: boolean;
-}
+export const CARDANO_NETWORKS = ['mainnet', 'testnet', 'preview', 'preprod'];
+
+export type Network = 'mainnet' | 'testnet' | 'preview' | 'preprod';
 
 export interface ResultFound {
   result: number;
 }
 
-export const CARDANO_NETWORKS = ['mainnet', 'testnet', 'preview', 'preprod'];
+// less generic types
+export interface ByronEraParameters {
+  epoch_length: number;
+  slot_length: number;
+  safe_zone: number;
+  end_epoch: number;
+}
 
-export type Network = 'mainnet' | 'testnet' | 'preview' | 'preprod';
+export interface ProtocolInfo {
+  name: string;
+  is_era_hardfork: boolean;
+}
+export interface Amount {
+  unit: string;
+  quantity: string;
+}
+
+type OnchainMetadataItem = components['schemas']['asset']['onchain_metadata'];
+
+export type OnchainMetadata = {
+  version?: number;
+} & Record<string, Record<string, OnchainMetadataItem>>;