+++ src/utils/routes.ts
@@ -1,0 +1,2 @@
+export const standardSafeZone = (securityParameter: number, activeSlotsCoeff: number) =>
+  (3 * securityParameter) / activeSlotsCoeff;