+++ test/unit/tests/utils/validation.ts
@@ -1,6 +1,4 @@
-import sinon from 'sinon';
 import { describe, expect, test, vi } from 'vitest';
-
 import * as config from '../../../../src/config';
 import * as validationUtils from '../../../../src/utils/validation';
 import {
@@ -14,7 +12,7 @@ import {
   validatePositiveInRangeSignedBigIntFixture,
   validatePositiveInRangeSignedIntFixture,
   validateStakeAddressFixture,
-} from '../../fixtures/validation-format-utils.fixtures';
+} from '../../fixtures/validation.fixtures';
 
 describe('validation-format-utils', () => {
   validateStakeAddressFixture.map(fixture => {
@@ -26,7 +24,7 @@ describe('validation-format-utils', () => {
 
       const result = validationUtils.validateStakeAddress(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -34,7 +32,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.validateAndConvertPool(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -42,7 +40,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.validateDerivationXpub(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -50,7 +48,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.validateInRangeUnsignedInt(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -58,7 +56,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.validatePositiveInRangeSignedInt(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -66,7 +64,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.validatePositiveInRangeSignedBigInt(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -74,7 +72,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.validateBlockHash(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -82,7 +80,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.isNumber(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -95,7 +93,7 @@ describe('validation-format-utils', () => {
 
       const result = validationUtils.convertStakeAddress(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 
@@ -103,7 +101,7 @@ describe('validation-format-utils', () => {
     test(fixture.name, async () => {
       const result = validationUtils.validateHex(fixture.input);
 
-      expect(result).toBe(fixture.response);
+      expect(result).toStrictEqual(fixture.response);
     });
   });
 });