+++ vitest.config.ts
@@ -2,6 +2,7 @@ import { defineConfig, defaultInclude } from 'vitest/config';
 
 export default defineConfig({
   test: {
+    testTimeout: 20_000,
     root: '.',
     mockReset: true,
     setupFiles: ['test/unit/setup.ts'],